fix(ui): More Advanced Filters Fixes
